Replace old VecBuffer with impl from VecBuffer2
