+++ src/sinks/aws_kinesis_firehose.rs
@@ -8,7 +8,7 @@ use crate::{
         rusoto,
         service2::TowerRequestConfig,
         sink::Response,
-        BatchConfig, BatchSettings, Compression, EncodedLength, VecBuffer2,
+        BatchConfig, BatchSettings, Compression, EncodedLength, VecBuffer,
     },
     topology::config::{DataType, SinkConfig, SinkContext, SinkDescription},
 };
@@ -143,7 +143,7 @@ impl KinesisFirehoseService {
             .batch_sink(
                 KinesisFirehoseRetryLogic,
                 kinesis,
-                VecBuffer2::new(batch.size),
+                VecBuffer::new(batch.size),
                 batch.timeout,
                 cx.acker(),
             )
