+++ src/sinks/loki.rs
@@ -19,7 +19,7 @@ use crate::{
         encoding::{EncodingConfigWithDefault, EncodingConfiguration},
         http::{Auth, BatchedHttpSink, HttpClient, HttpSink},
         service2::TowerRequestConfig,
-        BatchConfig, BatchSettings, BoxedRawValue, EncodedLength, UriSerde, VecBuffer2,
+        BatchConfig, BatchSettings, BoxedRawValue, EncodedLength, UriSerde, VecBuffer,
     },
     template::Template,
     tls::{TlsOptions, TlsSettings},
@@ -89,7 +89,7 @@ impl SinkConfig for LokiConfig {
 
         let sink = BatchedHttpSink::new(
             self.clone(),
-            VecBuffer2::new(batch_settings.size),
+            VecBuffer::new(batch_settings.size),
             request_settings,
             batch_settings.timeout,
             client.clone(),
