+++ src/sinks/util/buffer/mod.rs
@@ -9,7 +9,6 @@ pub mod json;
 pub mod metrics;
 pub mod partition;
 pub mod vec;
-pub mod vec2;
 
 pub use partition::{Partition, PartitionBuffer, PartitionInnerBuffer};
