+++ src/sinks/util/mod.rs
@@ -27,8 +27,7 @@ pub use batch::{Batch, BatchConfig, BatchSettings, BatchSize, PushResult};
 pub use buffer::json::{BoxedRawValue, JsonArrayBuffer};
 pub use buffer::metrics::{MetricBuffer, MetricEntry};
 pub use buffer::partition::Partition;
-pub use buffer::vec::VecBuffer;
-pub use buffer::vec2::{EncodedLength, VecBuffer2};
+pub use buffer::vec::{EncodedLength, VecBuffer};
 pub use buffer::{Buffer, Compression, PartitionBuffer, PartitionInnerBuffer};
 pub use service::{ServiceBuilderExt, TowerRequestConfig, TowerRequestLayer, TowerRequestSettings};
 pub use sink::{BatchSink, PartitionBatchSink, StreamSink};
